split and moved tests
